requirements.txt
SQLAlchemy==2.0.0
python-dotenv==1.0.0
psycopg2==2.9.6
requests==2.27.1
fastapi==0.109.2
uvicorn==0.22.0
schedule==1.1.0
aioredis==2.0.1
pydantic==2.5.0
alembic==1.11.1
aiogram==3.3.0
redis==4.4.1